Refactoring project as a Storm project
